Backed out changeset fe104053f4f2 (bug 861039) for compilation failures
UltraBlame original commit: 5a3487239596c1b0fe54cf46ff466cc2d9d45e35

+++ gfx/angle/src/compiler/MapLongVariableNames.h
@@ -31,7 +31,7 @@ public:
     void Insert(const char* originalName, const char* mappedName);
 
     
-    size_t Size() const;
+    int Size() const;
 
 private:
     LongNameMap();
